+++ trunk/src/l1j/jrwz/server/templates/L1EtcItem.java
@@ -30,7 +30,7 @@ public class L1EtcItem extends L1Item {
 
     private int _locy;
 
-    private short _mapid;
+    private int _mapid;
 
     private int _delay_id;
 
@@ -70,7 +70,7 @@ public class L1EtcItem extends L1Item {
     }
 
     @Override
-    public short get_mapid() {
+    public int get_mapid() {
         return _mapid;
     }
 
@@ -109,7 +109,7 @@ public class L1EtcItem extends L1Item {
         _locy = locy;
     }
 
-    public void set_mapid(short mapid) {
+    public void set_mapid(int mapid) {
         _mapid = mapid;
     }
